requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-docx>=1.1.0
python-multipart>=0.0.6
dashscope>=1.14.0
pydantic>=2.5.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
